restructure tests and dynatrace server handler to avoid duplication
